Alter price to float value

+++ backend/models/store_product_model.py
@@ -7,6 +7,6 @@ from .store_model import store
 class store_product(models.Model):
     store = models.ForeignKey(store, on_delete=models.CASCADE)
     product = models.ForeignKey(product, on_delete=models.CASCADE)
-    price  = models.IntegerField(default=0)
+    price  = models.FloatField(default=0)
     amount  = models.IntegerField(default=1)
